Rework docgen to use same example session for functions, tests, events and actions
