+++ cmd/docgen/templates/docs.md
@@ -393,19 +393,19 @@ Functions are called using the `@(function_name(args..))` syntax. For example, t
 Context variables referred to within functions do not need a leading `@`. Functions can also use literal numbers or strings as arguments, for example
 `@(length(split("1 2 3", " "))`.
 
-<div class="excellent_functions">
-{{ .ExcellentFunctionDocs }}
+<div class="functions">
+{{ .FunctionDocs }}
 </div>
 
-# Template Tests
+# Router Tests
 
-Template tests are a special class of functions which are used within the switch router. They are called in the same way as normal functions, but 
+Router tests are a special class of functions which are used within the switch router. They are called in the same way as normal functions, but 
 all return a test result object which by default evalutes to true or false, but can also be used to find the matching portion of the test by using
 the `match` component of the result. The flow editor builds these expressions using UI widgets, but they can be used anywhere a normal template
 function is used.
 
-<div class="excellent_tests">
-{{ .ExcellentTestDocs }}
+<div class="tests">
+{{ .TestDocs }}
 </div>
 
 # Action Definitions
