+++ cmd/docgen/templates/template.html
@@ -44,12 +44,12 @@
             border: 1px solid rgb(107, 79, 17);
         }
 
-        .excellent_functions h2 {
+        .functions h2 {
             background-color: #2B823A;
             border: 1px solid #10621E;
         }
 
-        .excellent_tests h2 {
+        .tests h2 {
             background-color: #8b257e;
             border: 1px solid rgb(78, 14, 64);
         }
